=> Confirmed that all the VM tests are passing
=> Updated memory to use uint64_t rather than size_t 

=> Fixed a bug in memory, where it was possible to read a slice out of bounds 

=> Fixed the misspelt reference to ext in eos_execute 

=> Fixed a bug where the ret and arg of the DUP9+ instructions were using base10 rather than hex 

=> Fixed a bug where the ret and arg of the SWAP9+ instructions were using base10 rather than hex

+++ src/evm/include/evm/types.h
@@ -44,7 +44,7 @@ enum ActionType {
 
 typedef ActionType action_type_t;
 
-typedef int gas_t;
+typedef uint64_t gas_t;
 
 struct Params {
   uint256_t codeAddress;
